requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow==10.0.1
python-dotenv==1.0.0
requests==2.31.0
openai>=1.0.0
dashscope>=1.23.8